Find mass of body either on Earth or on Moon
Weight is defined as the gravitational force acting on the mass of a body. The weight of a body with a mass of 1kg on Earth can be calculated with the formula given below. The weight is expressed in Newtons (N).
          WEIGHT ON EARTH = (1 Kg ) * ( 9.81 m/s2 ) = 9.81 N
The weight of the same body in N can be calculated with another formula:
         WEIGHT ON MOON = ((1 Kg ) * ( 9.81 m/s2)) /6 = 1.64 N
Write a program that takes as input the mass of a car in kilograms and then prompts the user to choose to calculate the car's weight in Newton either on Earth or on the Moon. The program should allow the user to enter new values and output  the car's weight on Earth or on the Moon till the user decides to exit. user the double data type to store the mass.